requirements.txt
fastapi
uvicorn
langchain-google-genai
langchain
python-dotenv
langchain_community
tavily-python